Generalize trainer to be reused
